+++ R/zzz.R
@@ -14,6 +14,6 @@
                              "cname", "first_num", "new_level_key", "zs", "dfn", 
                              "cnt", "col_names", "white_labs", "black_labs", 
                              "cls", "alpha", "jsd", "significant", "bar_width", 
-                             "corr_1", "corr_2", "data_frame_n"))
+                             "corr_1", "corr_2", "data_frame_n", "col_vec", "size"))
   invisible()
 }